Move doc generator stream to lib/streams
The test file too
