+++ lib/streams/peliasDocGenerator.js
